Move graphql schema stuff into schema/
